services/waha-integrator/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.4
pydantic==2.8.2
pydantic-settings==2.4.0
python-dotenv==1.0.1
httpx==0.27.0